- Added INTERNAL flag (https://issues.jboss.org/browse/JGRP-1599) - Added internal thread pool to TP (https://issues.jboss.org/browse/JGRP-1599) - Replaced Message.OOB with Message.Flag.OOB - Replaced Message.DONT_BUNDLE with Message.Flag.DONT_BUNDLE - Failure detection protocols now use INTERNAL flag - Flow control protocols now use INTERNAL flag - Membership protocols now use INTERNAL flag - Stability protocol now uses INTERNAL flag - MERGE3 protocol now uses INTERNAL flag - Reliable retransmission protocols now use INTERNAL flag - Sending discovery messages with DONT_BUNDLE (required)

+++ src/org/jgroups/protocols/UDP.java
@@ -603,12 +603,12 @@ public class UDP extends TP {
         if(map == null) return;
 
         if(map.containsKey("send_buf_size")) {
-            mcast_send_buf_size=((Integer)map.get("send_buf_size")).intValue();
+            mcast_send_buf_size=(Integer)map.get("send_buf_size");
             ucast_send_buf_size=mcast_send_buf_size;
             set_buffers=true;
         }
         if(map.containsKey("recv_buf_size")) {
-            mcast_recv_buf_size=((Integer)map.get("recv_buf_size")).intValue();
+            mcast_recv_buf_size=(Integer)map.get("recv_buf_size");
             ucast_recv_buf_size=mcast_recv_buf_size;
             set_buffers=true;
         }
